Add a new chart type for drawing and handling context changes.
The new chart listens for changes in the context value or operator and
updates the PlotBands and PlotLines on the chart accordingly. Not all of
the operators and currently supported and I haven't confirmed this
visually but the new chart test is working so I'll consider this working
until the field controls(drop downs and text boxes) are working and
modifying the context.