Update to Bootstrap 3
This replaces the Google-themed variant for the default flat style of
BS3. Font-Awesome has been updated to 3.2.1.

The SCSS files have been removed and added as normal CSS files. This enables
use of those files without throughout a project without needing to load
Cilantro's styles when not applicable. This only speeds up the sass
compilation since the heavy dependencies are no longer included.